requirements.txt
torch>=2.1.0
diffusers>=0.27.0
transformers>=4.40.0
accelerate>=0.30.0
safetensors>=0.4.2
Pillow>=10.3.0
opencv-python>=4.8.1.78
tqdm>=4.66.4